// src/types/SmartSearchOptions.ts
export interface SmartSearchOptions {
  maxDropDownHeight?: number
  maxMatcherWidth?: number
  showCategories?: boolean
  categoryPosition?: 'top' | 'left'
  hideToolTip?: boolean
  promiseDelay?: number
  mockPromiseTime?: number
  defaultItemLimit?: number
  showWhenSearching?: boolean
  hideHelp?: boolean
}